+++ canvas_sdk/test_utils/factories/__init__.py
@@ -1,4 +1,5 @@
 from .calendar import CalendarFactory, EventFactory
+from .change_medication import ChangeMedicationFactory
 from .chart_section_review import ChartSectionReviewFactory
 from .claim import (
     ClaimCommentFactory,
@@ -21,6 +22,7 @@ from .claim_line_item import (
 from .coverage import CoverageFactory
 from .diagnostic_report import DiagnosticReportFactory
 from .django_content_type import ContentTypeFactory
+from .document_review_delegation import DocumentReviewDelegationFactory
 from .external_event import ExternalEventFactory, ExternalVisitFactory
 from .facility import FacilityFactory
 from .imaging import (
@@ -54,11 +56,13 @@ from .lab_report_template import (
     LabReportTemplateFieldOptionFactory,
 )
 from .letter import LanguageFactory, LetterActionEventFactory, LetterFactory
+from .medication import MedicationFactory
 from .medication_history import (
     MedicationHistoryMedicationCodingFactory,
     MedicationHistoryMedicationFactory,
     MedicationHistoryResponseFactory,
 )
+from .medication_statement import MedicationStatementFactory
 from .note import NoteFactory, NoteMetadataFactory, NoteStateChangeEventFactory, NoteTypeFactory
 from .organization import (
     OrganizationAddressFactory,
@@ -116,6 +120,7 @@ from .visual_exam_finding import VisualExamFindingFactory
 __all__ = (
     "CalendarFactory",
     "CanvasUserFactory",
+    "ChangeMedicationFactory",
     "ChartSectionReviewFactory",
     "ClaimBannerAlertFactory",
     "ClaimFactory",
@@ -134,6 +139,7 @@ __all__ = (
     "CoverageFactory",
     "DiagnosticReportFactory",
     "ContentTypeFactory",
+    "DocumentReviewDelegationFactory",
     "EventFactory",
     "ExternalEventFactory",
     "ExternalVisitFactory",
@@ -166,9 +172,11 @@ __all__ = (
     "LabReportTemplateFieldFactory",
     "LabReportTemplateFieldOptionFactory",
     "LetterActionEventFactory",
+    "MedicationFactory",
     "MedicationHistoryMedicationFactory",
     "MedicationHistoryMedicationCodingFactory",
     "MedicationHistoryResponseFactory",
+    "MedicationStatementFactory",
     "NoteFactory",
     "NoteMetadataFactory",
     "NoteStateChangeEventFactory",