+++ canvas_sdk/v1/data/__init__.py
@@ -21,6 +21,7 @@ from .billing import BillingLineItem, BillingLineItemModifier
 from .business_line import BusinessLine
 from .calendar import Calendar, Event
 from .care_team import CareTeamMembership, CareTeamRole
+from .change_medication import ChangeMedication
 from .charge_description_master import ChargeDescriptionMaster
 from .chart_section_review import ChartSectionReview, ChartSectionReviewSection
 from .claim import (
@@ -60,6 +61,7 @@ from .document_reference import (
     DocumentReferenceCoding,
     DocumentReferenceStatus,
 )
+from .document_review_delegation import DocumentReviewDelegation
 from .encounter import Encounter
 from .external_event import ExternalEvent, ExternalVisit
 from .facility import Facility
@@ -180,6 +182,7 @@ from .questionnaire import (
 )
 from .reason_for_visit import ReasonForVisitSettingCoding
 from .referral import Referral, ReferralReport, ReferralReportCoding, ReferralReview
+from .refill_request import RefillRequest, RefillRequestCoding
 from .service_provider import ServiceProvider
 from .snapshot import Snapshot, SnapshotImage
 from .specialty_report_template import (
@@ -228,6 +231,7 @@ __all__ = __exports__ = (
     "CanvasUser",
     "CareTeamMembership",
     "CareTeamRole",
+    "ChangeMedication",
     "ChargeDescriptionMaster",
     "ChartSectionReview",
     "ChartSectionReviewSection",
@@ -272,6 +276,7 @@ __all__ = __exports__ = (
     "DocumentReferenceCategory",
     "DocumentReferenceCoding",
     "DocumentReferenceStatus",
+    "DocumentReviewDelegation",
     "EligibilitySummary",
     "Encounter",
     "Event",
@@ -388,6 +393,8 @@ __all__ = __exports__ = (
     "ReferralReport",
     "ReferralReportCoding",
     "ReferralReview",
+    "RefillRequest",
+    "RefillRequestCoding",
     "ResponseOption",
     "ResponseOptionSet",
     "ServiceProvider",